requirements.txt
aiogram
aiogram_dialog
aiogram_i18n
python-dotenv
motor
pymongo
fluent
fluent.runtime
libcst
betterlogging
